+++ src/dedicated_room/citra-room.cpp
@@ -178,7 +178,7 @@ int main(int argc, char** argv) {
     std::string ban_list_file;
     std::string log_file = "citra-room.log";
     u64 preferred_game_id = 0;
-    u32 port = Network::DefaultRoomPort;
+    u16 port = Network::DefaultRoomPort;
     u32 max_members = 16;
     bool enable_citra_mods = false;
 
@@ -212,7 +212,7 @@ int main(int argc, char** argv) {
                 room_description.assign(optarg);
                 break;
             case 'p':
-                port = strtoul(optarg, &endarg, 0);
+                port = static_cast<u16>(strtoul(optarg, &endarg, 0));
                 break;
             case 'm':
                 max_members = strtoul(optarg, &endarg, 0);
